[chore] repository key 타입 변경

+++ src/main/java/org/asansocketserver/domain/position/repository/BeaconDataRepository.java
@@ -8,16 +8,16 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.HashMap;
 import java.util.List;
 
-public interface BeaconDataRepository extends JpaRepository<BeaconData, Long> {
+public interface BeaconDataRepository extends JpaRepository<BeaconData, Integer> {
     List<BeaconData> findAllByPosition(String positionName);
 
-
-//    @Query("SELECT b.position, c.longitude, c.latitude, COUNT(b) FROM Coordinate c LEFT JOIN BeaconData b ON b.position = c.position GROUP BY b.position ,  c.longitude, c.latitude")
-//    List<Object[]> findAllBeaconCount();
+    // @Query("SELECT b.position, c.longitude, c.latitude, COUNT(b) FROM Coordinate
+    // c LEFT JOIN BeaconData b ON b.position = c.position GROUP BY b.position ,
+    // c.longitude, c.latitude")
+    // List<Object[]> findAllBeaconCount();
 
     @Query("SELECT c.position, COUNT(b) FROM Coordinate c LEFT JOIN BeaconData b ON b.position = c.position GROUP BY c.position")
     List<Object[]> findAllBeaconCount();
-
 
     public void deleteAllByImageId(Long imageId);
